backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "influxdb3-python>=0.10.0",
    "pydantic>=2.12.5",
    "snowflake-id>=1.0.2",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
